Module admission_control_proto::proto::admission_control

Modules

admission_control_client
Generated client implementations.

admission_control_msg

admission_control_server
Generated server implementations.

submit_transaction_response

Structs

AdmissionControlMsg
The request for submitting a transaction to an upstream validator or full node.

AdmissionControlStatus
AC response status containing code and optionally an error message.

SubmitTransactionRequest

SubmitTransactionResponse
The response for transaction submission.

Enums

AdmissionControlStatusCode
Additional statuses that are possible from admission control in addition to VM statuses.