Trait admission_control_proto::proto::admission_control::admission_control_server::AdmissionControl

pub trait AdmissionControl: Send + Sync + 'static {
#[must_use]
    fn submit_transaction<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<SubmitTransactionRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<SubmitTransactionResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn update_to_latest_ledger<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateToLatestLedgerRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<UpdateToLatestLedgerResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with AdmissionControlServer.

Required methods

#[must_use] fn submit_transaction<'life0, 'async_trait>(
&'life0 self,
request: Request<SubmitTransactionRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<SubmitTransactionResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Public API to submit transaction to a validator.

#[must_use] fn update_to_latest_ledger<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateToLatestLedgerRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<UpdateToLatestLedgerResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

This API is used to update the client to the latest ledger version and optionally also request 1..n other pieces of data. This allows for batch queries. All queries return proofs that a client should check to validate the data. Note that if a client only wishes to update to the latest LedgerInfo and receive the proof of this latest version, they can simply omit the requested_items (or pass an empty list)

Loading content...